% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\name{document_data}
\alias{document_data}
\title{Document data}
\usage{
document_data(x)
}
\arguments{
\item{x}{the name of data set to document. The dataset
must be load in the R environment.}
}
\value{
an R file named as same as the dataset.
}
\description{
Create an .R file containing information gathered from
the data set to document. The variables' names are inserted
in a roxygen template to facilitate the documentation process,
and the row and column numbers are added to the description.

}
\details{
document_data


}
